+++ front-end/src/UI/Templates/CreateChannel/index.tsx
@@ -10,7 +10,7 @@ import {
   FormikValues,
 } from 'formik';
 import * as Yup from 'yup';
-import WarningAlertDialog from '../WarningAlertDialog';
+import useWarningDialog from '../../../Hooks/useWarningDialog';
 
 const CreateChannelScheme = Yup.object().shape({
   name: Yup.string()
@@ -30,10 +30,7 @@ const CreateChannelScheme = Yup.object().shape({
 });
 
 function CreateChannel() {
-  const [error, setError] = React.useState({
-    headerMessage: '',
-    bodyMessage: '',
-  });
+  const { setError, WarningDialogComponent } = useWarningDialog();
   const onSubmitHandler = React.useCallback(
     (values: FormikValues, actions: FormikHelpers<FormikValues>) => {
       setTimeout(() => {
@@ -44,9 +41,8 @@ function CreateChannel() {
         actions.setSubmitting(false);
       }, 1000);
     },
-    [],
+    [setError],
   );
-  const cancelRef = React.useRef(null); // TODO: 삭제하기
 
   return (
     <Formik
@@ -111,13 +107,7 @@ function CreateChannel() {
               </GridItem>
             </Grid>
           </Form>
-          <WarningAlertDialog
-            isOpen={error.bodyMessage.length > 0}
-            onClose={() => setError({ headerMessage: '', bodyMessage: '' })}
-            cancelRef={cancelRef}
-            headerMessage={error.headerMessage}
-            bodyMessage={error.bodyMessage}
-          />
+          {WarningDialogComponent}
         </>
       )}
     </Formik>
